# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ScheduleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"repeatType": {
						"type": "string"
					},
					"notification": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"userId",
					"description",
					"startDate",
					"endDate",
					"repeatType",
					"notification"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ScheduleResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ScheduleResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorDetails": {
				"anyOf": [
					{
						"properties": {
							"scheduleId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"reason": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"endDate": {
								"type": "string",
								"format": "date-time"
							},
							"startDate": {
								"type": "string",
								"format": "date-time"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"groupId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"userId": {
								"type": "string"
							},
							"groupId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"focusTargetId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				],
				"nullable": true
			},
			"ITsoaErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ErrorDetails"
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BodyToSchedule": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"repeatType": {
						"type": "string"
					},
					"notification": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"description",
					"startDate",
					"endDate",
					"repeatType",
					"notification"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"sessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			}
		}
	},
	"info": {
		"title": "Server",
		"version": "1.0.0",
		"description": "25-1 Capstone Server",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "asjasj3964 ",
			"email": "asjasj3964@naver.com"
		}
	},
	"paths": {
		"/schedule": {
			"post": {
				"operationId": "HandleScheduleAdd",
				"responses": {
					"200": {
						"description": "일정 생성 성공 응답",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ScheduleResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"id": "1",
												"title": "string",
												"userId": "1",
												"description": "string",
												"startDate": "2025-01-17T03:50:25",
												"endDate": "2025-01-17T03:50:25",
												"repeatType": "반복 없음",
												"notification": 0
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "유효하지 않은 데이터 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SCH-400",
												"reason": "일정 생성 중 오류가 발생했습니다.",
												"data": {
													"scheduleId": "1"
												}
											}
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SCH-400",
												"reason": "종료일을 시작일과 일치하거나 과거로 설정할 수 없습니다.",
												"data": {
													"startDate": "2025-05-16T18:50:25.000Z",
													"endDate": "2025-05-12T18:50:25.000Z"
												}
											}
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SCH-400",
												"reason": "이미 존재하는 일정과 시간이 겹칩니다.",
												"data": {
													"startDate": "2025-05-16T18:50:25.000Z",
													"endDate": "2025-05-17T18:50:25.000Z"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "일정을 생성하는 API입니다.",
				"summary": "일정 생성 API",
				"tags": [
					"Schedule-Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "일정 제목, 설명, 시작일, 종료일, 반복 여부, 알림",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BodyToSchedule",
								"description": "일정 제목, 설명, 시작일, 종료일, 반복 여부, 알림"
							}
						}
					}
				}
			}
		},
...
		"/schedule/{scheduleId}": {
			"get": {
				"operationId": "HandleScheduleGet",
				"responses": {
					"200": {
						"description": "특정 일정 조회 성공 응답",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ScheduleResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"userId": "1",
												"schedules": {
													"id": "1",
													"title": "string",
													"description": "string",
													"startDate": "2025-01-17T03:50:25",
													"endDate": "2025-01-17T03:50:25",
													"repeatType": "반복 안함",
													"notification": 0
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않은 데이터 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SCH-404",
												"reason": "해당 일정을 찾을 수 없습니다.",
												"data": {
													"scheduleId": "1"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 일정을 조회하는 API입니다.",
				"summary": "특정 일정 조회 API",
				"tags": [
					"Schedule-Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "scheduleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://ai-study-lens.p-e.kr",
			"description": "Production Server"
		},
		{
			"url": "http://localhost:3000",
			"description": "local server"
		}
	]
}